--- apps/web/components/NewsCardNew/index.ts
@@ -1,1 +1,0 @@
-export { default as NewsCardNew } from './NewsCardNew'
