+++ apps/web/components/real.ts
@@ -29,7 +29,6 @@ export * from './ContentLink/ContentLink'
 export * from './LatestNewsSection'
 export * from './LatestNewsSectionSlider'
 export * from './NewsCard'
-export * from './NewsCardNew'
 export * from './LogoList/LogoList'
 export * from './List/List'
 export * from './AboutLatestNews/AboutLatestNews'
